+++ resources/hpxml-measures/HPXMLtoOpenStudio/resources/hvac_sizing.rb
@@ -441,7 +441,7 @@ module HVACSizing
                 case roof.roof_color
                 when HPXML::ColorDark, HPXML::ColorMediumDark
                   cool_temp += 130.0 * roof.net_area
-                when HPXML::ColorMedium, HPXML::ColorLight
+                when HPXML::ColorMedium, HPXML::ColorMediumLight, HPXML::ColorLight
                   cool_temp += 120.0 * roof.net_area
                 when HPXML::ColorReflective
                   cool_temp += 95.0 * roof.net_area
@@ -450,7 +450,7 @@ module HVACSizing
                 case roof.roof_color
                 when HPXML::ColorDark, HPXML::ColorMediumDark
                   cool_temp += 110.0 * roof.net_area
-                when HPXML::ColorMedium, HPXML::ColorLight
+                when HPXML::ColorMedium, HPXML::ColorMediumLight, HPXML::ColorLight
                   cool_temp += 105.0 * roof.net_area
                 when HPXML::ColorReflective
                   cool_temp += 95.0 * roof.net_area
@@ -471,7 +471,7 @@ module HVACSizing
                 case roof.roof_color
                 when HPXML::ColorDark, HPXML::ColorMediumDark
                   cool_temp += 120.0 * roof.net_area
-                when HPXML::ColorMedium, HPXML::ColorLight
+                when HPXML::ColorMedium, HPXML::ColorMediumLight, HPXML::ColorLight
                   cool_temp += 110.0 * roof.net_area
                 when HPXML::ColorReflective
                   cool_temp += 95.0 * roof.net_area
@@ -480,7 +480,7 @@ module HVACSizing
                 case roof.roof_color
                 when HPXML::ColorDark, HPXML::ColorMediumDark
                   cool_temp += 105.0 * roof.net_area
-                when HPXML::ColorMedium, HPXML::ColorLight
+                when HPXML::ColorMedium, HPXML::ColorMediumLight, HPXML::ColorLight
                   cool_temp += 100.0 * roof.net_area
                 when HPXML::ColorReflective
                   cool_temp += 95.0 * roof.net_area
@@ -1171,7 +1171,7 @@ module HVACSizing
         if [HPXML::RoofTypeClayTile, HPXML::RoofTypeWoodShingles].include? roof.roof_type
           cltd *= 0.83
         end
-      when HPXML::ColorMedium, HPXML::ColorLight
+      when HPXML::ColorMedium, HPXML::ColorMediumLight, HPXML::ColorLight
         if [HPXML::RoofTypeClayTile].include? roof.roof_type
           cltd *= 0.65
         else
@@ -2870,6 +2870,8 @@ module HVACSizing
       end
     end
 
+    heating_load = hvac_sizings.Heat_Load
+
     if hvac_sizings.Heat_Load <= 0
 
       hvac_sizings.Heat_Capacity = 0.0
@@ -2881,7 +2883,7 @@ module HVACSizing
            HPXML::HVACTypeHeatPumpPTHP,
            HPXML::HVACTypeHeatPumpRoom].include? heating_type
 
-      calculate_heat_pump_capacities(mj, runner, hvac_sizings, weather, hvac_heating, total_cap_curve_value, hvac_system, oversize_limit, oversize_delta, hpxml_bldg, hpxml_header)
+      heating_load = calculate_heat_pump_capacities(mj, runner, hvac_sizings, weather, hvac_heating, total_cap_curve_value, hvac_system, oversize_limit, oversize_delta, hpxml_bldg, hpxml_header)
 
       hvac_sizings.Heat_Capacity_Supp = calculate_heat_pump_backup_load(mj, hvac_heating, hvac_sizings.Heat_Load_Supp, hvac_sizings.Heat_Capacity, hpxml_bldg)
       hvac_sizings.Heat_Airflow = calc_airflow_rate(:htg, hvac_heating, hvac_sizings.Heat_Capacity, hpxml_bldg)
@@ -2953,7 +2955,7 @@ module HVACSizing
     # If HERS sizing methodology, ensure HP capacity is at least equal to larger of
     # heating and sensible cooling loads.
     if is_heatpump_with_both_htg_and_clg && (hpxml_bldg.header.heat_pump_sizing_methodology == HPXML::HeatPumpSizingHERS)
-      min_capacity = [hvac_sizings.Heat_Load, hvac_sizings.Cool_Load_Sens].max
+      min_capacity = [heating_load, hvac_sizings.Cool_Load_Sens].max
       if hvac_sizings.Cool_Capacity < min_capacity
         scaling_factor = min_capacity / hvac_sizings.Cool_Capacity
         hvac_sizings.Cool_Capacity *= scaling_factor
@@ -3641,7 +3643,7 @@ module HVACSizing
   # @param oversize_limit [Double] Oversize fraction (frac)
   # @param oversize_delta [Double] Oversize delta (Btu/hr)
   # @param hpxml_bldg [HPXML::Building] HPXML Building object representing an individual dwelling unit
-  # @return [nil]
+  # @return [Double] Design heating load used for sizing the heat pump
   def self.calculate_heat_pump_capacities(mj, runner, hvac_sizings, weather, hvac_heating, cool_cap_adj_factor, hvac_system,
                                           oversize_limit, oversize_delta, hpxml_bldg, hpxml_header)
 
@@ -3697,6 +3699,8 @@ module HVACSizing
       hvac_sizings.Cool_Airflow = cfm_per_btuh * hvac_sizings.Cool_Capacity
       hvac_sizings.Heat_Capacity = hvac_sizings.Cool_Capacity
     end
+
+    return heating_load
   end
 
   # Retrieves a collection of ventilation information from the HPXML building.
@@ -4422,7 +4426,7 @@ module HVACSizing
       wall_ufactor = 1.0 / wall.insulation_assembly_r_value
 
       if wall_type == HPXML::WallTypeWoodStud
-        if wall.siding == HPXML::SidingTypeBrick
+        if [HPXML::SidingTypeBrick, HPXML::SidingTypeStone].include? wall.siding
           if wall_ufactor <= 0.070
             table_4a_wall_group = 'K'
           elsif wall_ufactor <= 0.083
@@ -4463,7 +4467,7 @@ module HVACSizing
         end
 
       elsif wall_type == HPXML::WallTypeSteelStud
-        if wall.siding == HPXML::SidingTypeBrick
+        if [HPXML::SidingTypeBrick, HPXML::SidingTypeStone].include? wall.siding
           if wall_ufactor <= 0.090
             table_4a_wall_group = 'K'
           elsif wall_ufactor <= 0.105
@@ -4505,20 +4509,20 @@ module HVACSizing
 
       elsif wall_type == HPXML::WallTypeDoubleWoodStud
         table_4a_wall_group = 'J' # assumed since MJ8 does not include double stud constructions
-        if wall.siding == HPXML::SidingTypeBrick
+        if [HPXML::SidingTypeBrick, HPXML::SidingTypeStone].include? wall.siding
           table_4a_wall_group = 'K'
         end
 
       elsif wall_type == HPXML::WallTypeSIP
         # Manual J refers to SIPs as Structural Foam Panel (SFP)
         if wall_ufactor >= (0.072 + 0.050) / 2
-          if wall.siding == HPXML::SidingTypeBrick
+          if [HPXML::SidingTypeBrick, HPXML::SidingTypeStone].include? wall.siding
             table_4a_wall_group = 'J'
           else
             table_4a_wall_group = 'G'
           end
         elsif wall_ufactor >= 0.050
-          if wall.siding == HPXML::SidingTypeBrick
+          if [HPXML::SidingTypeBrick, HPXML::SidingTypeStone].include? wall.siding
             table_4a_wall_group = 'K'
           else
             table_4a_wall_group = 'I'
@@ -4543,7 +4547,7 @@ module HVACSizing
           table_4a_wall_group = 'E'
         end
 
-      elsif [HPXML::WallTypeBrick, HPXML::WallTypeAdobe, HPXML::WallTypeConcrete].include? wall_type
+      elsif [HPXML::WallTypeBrick, HPXML::WallTypeAdobe, HPXML::WallTypeConcrete, HPXML::WallTypeStone].include? wall_type
         # Two Courses Brick or 8 Inches Concrete
         if wall_ufactor >= (0.218 + 0.179) / 2
           table_4a_wall_group = 'G'
@@ -4571,7 +4575,7 @@ module HVACSizing
           table_4a_wall_group = 'K'
         end
 
-      elsif [HPXML::WallTypeICF, HPXML::WallTypeStrawBale, HPXML::WallTypeStone].include? wall_type
+      elsif [HPXML::WallTypeICF, HPXML::WallTypeStrawBale].include? wall_type
         table_4a_wall_group = 'K'
 
       end
@@ -4952,7 +4956,7 @@ module HVACSizing
   # @param window [HPXML::Window] The window of interest
   # @return [Double] Interior shading coefficient
   def self.get_window_interior_shading_coefficient(window)
-    return 1.0 if window.interior_shading_type == HPXML::InteriorShadingTypeNone
+    return 1.0 if window.interior_shading_type == HPXML::InteriorShadingTypeNotPresent
 
     # Look up window type
     if window.glass_layers.nil? || window.glass_type.nil?