+++ resources/hpxml-measures/workflow/tests/util.rb
@@ -364,7 +364,6 @@ def _verify_outputs(rundir, hpxml_path, results, hpxml, unit_multiplier)
     next if message.include?('setupIHGOutputs: Output variables=Zone Other Equipment') && message.include?('are not available.')
     next if message.include?('setupIHGOutputs: Output variables=Space Other Equipment') && message.include?('are not available')
     next if message.include? 'Multiple speed fan will be applied to this unit. The speed number is determined by load.'
-    next if message.include?('Temperature') && message.include?('out of bounds') && message.include?('ATTIC')
 
     # HPWHs
     if hpxml.buildings.any? { |hpxml_bldg| hpxml_bldg.water_heating_systems.count { |wh| wh.water_heater_type == HPXML::WaterHeaterTypeHeatPump } > 0 }
@@ -1081,6 +1080,11 @@ def _verify_outputs(rundir, hpxml_path, results, hpxml, unit_multiplier)
   end
 
   # Check unmet hours
+  skip_unmet_check = false
+  if hpxml_path.include?('install-quality') || hpxml_path.include?('research-features')
+    # unmet hours are expected for HVAC installation quality and realistic backup staging files
+    skip_unmet_check = true
+  end
   unmet_hours_htg = results.select { |k, _v| k.include? 'Unmet Hours: Heating' }.values.sum(0.0)
   unmet_hours_clg = results.select { |k, _v| k.include? 'Unmet Hours: Cooling' }.values.sum(0.0)
   if hpxml_path.include? 'base-hvac-undersized.xml'
@@ -1090,13 +1094,12 @@ def _verify_outputs(rundir, hpxml_path, results, hpxml, unit_multiplier)
     if hpxml_bldg.total_fraction_heat_load_served == 0
       assert_equal(0, unmet_hours_htg)
     else
-      # for realistic backup staging, unmet hours are expected.
-      assert_operator(unmet_hours_htg, :<, 500) unless hpxml_path.include? 'research-features'
+      assert_operator(unmet_hours_htg, :<, 500) unless skip_unmet_check
     end
     if hpxml_bldg.total_fraction_cool_load_served == 0
       assert_equal(0, unmet_hours_clg)
     else
-      assert_operator(unmet_hours_clg, :<, 500)
+      assert_operator(unmet_hours_clg, :<, 500) unless skip_unmet_check
     end
   end
 
@@ -1132,9 +1135,9 @@ def _check_unit_multiplier_results(xml, hpxml_bldg, annual_results_1x, annual_re
       abs_delta_tol = 500.0
       abs_frac_tol = 0.15
     elsif key.include?('Peak Load:')
-      # Check that the peak load difference is less than 0.2 kBtu/hr or less than 5%
+      # Check that the peak load difference is less than 0.2 kBtu/hr or less than 10%
       abs_delta_tol = 0.2
-      abs_frac_tol = 0.05
+      abs_frac_tol = 0.1
     elsif key.include?('Hot Water:')
       # Check that the hot water usage difference is less than 10 gal/yr or less than 2%
       abs_delta_tol = 10.0
@@ -1144,7 +1147,7 @@ def _check_unit_multiplier_results(xml, hpxml_bldg, annual_results_1x, annual_re
       abs_delta_tol = 1.0
       abs_frac_tol = 0.01
     elsif key.include?('Airflow:')
-      # Check that airflow rate difference is less than 0.2 cfm or less than 5.0%
+      # Check that airflow rate difference is less than 0.2 cfm or less than 5%
       abs_delta_tol = 0.2
       abs_frac_tol = 0.05
     elsif key.include?('Unmet Hours:')